Switch assertion failure to returning failure
This assertion failure can be hit in production, which causes the
openvpn server process to stop and all clients to be disconnected.

Tracing callers, it appears that some callers check ks->authenticated
before calling, but others do not.  It may be possible to add the check
for the callers that do not check, but this seems to be a similar
solution.